Change bottom of the main page button on hover, change some font styles, move tests to the test folder
